Rename 'redux' dir to 'store'
* This is to avoid conflicts with 'redux' imports later as we use absolute imports from 'src'
