+++ client/src/app/header.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { common } from '../common.styles';
-import { actions } from '../redux/actions';
-import { State } from '../redux/reducers';
+import { actions } from '../store/actions';
+import { State } from '../store/reducers';
 import pick from 'lodash/pick';
 import { styles } from './header.styles';
