+++ client/src/components/globalSpinner.tsx
@@ -1,8 +1,8 @@
 import { css } from 'aphrodite/no-important';
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { State } from '../redux/reducers';
-import * as selectors from '../redux/selectors';
+import { State } from '../store/reducers';
+import * as selectors from '../store/selectors';
 import { FadeIn } from './animations';
 import { styles } from './globalSpinner.styles';
