+++ client/src/hocs/requireUserLogin.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { actions } from '../redux/actions';
-import { State } from '../redux/reducers';
-import * as selectors from '../redux/selectors';
+import { actions } from '../store/actions';
+import { State } from '../store/reducers';
+import * as selectors from '../store/selectors';
 import pick from 'lodash/pick';
 
 interface IProps {
