+++ client/src/pages/addNotablePerson/addNotablePerson.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { State } from '../../redux/reducers';
-import * as selectors from '../../redux/selectors';
+import { State } from '../../store/reducers';
+import * as selectors from '../../store/selectors';
 import pick from 'lodash/pick';
 
 interface IProps {
